feat: add error messages
